pyproject.toml
[project]
name = "gemini-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.16.1",
    "mcp>=1.9.0",
]
